Add files via upload
We are actually closer to the correct outcome than the previous code we ported from though the iterations take longer because we are printing out so much data for feedback from the neuron as one training set. If anything we are slightly over the mark. However, to reduce the feedback and get the output feeding to multiple neurons will be exciting. sampe_data.txt